+++ emcc.py
@@ -1176,8 +1176,6 @@ def run(args):
   # Strip args[0] (program name)
   args = args[1:]
 
-  misc_temp_files = shared.get_temp_files()
-
   # Handle some global flags
 
   # read response files very early on
@@ -1223,7 +1221,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
   if '--cflags' in args:
     # fake running the command, to see the full args we pass to clang
     args = [x for x in args if x != '--cflags']
-    with misc_temp_files.get_file(suffix='.o') as temp_target:
+    with shared.get_temp_files().get_file(suffix='.o') as temp_target:
       input_file = 'hello_world.c'
       compiler = shared.EMCC
       if run_via_emxx:
@@ -3241,17 +3239,23 @@ def phase_embind_emit_tsd(options, in_wasm, wasm_target, memfile, js_syms):
   # Required function to trigger TS generation.
   settings.DEFAULT_LIBRARY_FUNCS_TO_INCLUDE += ['$callRuntimeCallbacks']
   settings.EXPORT_ES6 = False
+  # Disable minify since the binaryen pass has not been run yet to change the
+  # import names.
+  settings.MINIFY_WASM_IMPORTED_MODULES = False
   setup_environment_settings()
+  # Embind may be included multiple times, de-duplicate the list first.
+  settings.JS_LIBRARIES = dedup_list(settings.JS_LIBRARIES)
   # Replace embind with the TypeScript generation version.
   embind_index = settings.JS_LIBRARIES.index('embind/embind.js')
   settings.JS_LIBRARIES[embind_index] = 'embind/embind_ts.js'
 
-  outfile_js = os.path.join(os.path.dirname(wasm_target), 'tsgen_a.out.js')
-  emscripten.run(in_wasm, wasm_target, outfile_js, memfile, js_syms)
+  outfile_js = in_temp('tsgen_a.out.js')
+  # The Wasm outfile may be modified by emscripten.run, so use a temporary file.
+  outfile_wasm = in_temp('tsgen_a.out.wasm')
+  emscripten.run(in_wasm, outfile_wasm, outfile_js, memfile, js_syms)
   out = shared.run_js_tool(outfile_js, [], stdout=PIPE)
   write_file(
     os.path.join(os.path.dirname(wasm_target), options.embind_emit_tsd), out)
-  delete_file(outfile_js)
   settings.restore(original_settings)