+++ ChangeLog.md
@@ -20,6 +20,7 @@ See docs/process.md for more on how version tagging works.
 
 3.1.47 (in development)
 -----------------------
+- The boost-headers port was updated to v1.83.0.
 - The `noExitRuntime` global is now a JS library symbol that will only be
   included as needed.  User of `noExitRuntime` will now need to declare a
   dependency on it.  It will get implicitly included if `noExitRuntime` is part
@@ -39,6 +40,8 @@ See docs/process.md for more on how version tagging works.
 - `emscripten::val` now prevents accidental access to the underlying JavaScript
   value from threads other than its owner. This already didn't work correctly
   in majority of cases, but now it will throw a clear assertion failure. (#20344)
+- `emscripten::val` can now be iterated over with a C++ range-based for loop.
+  (#20364)
 
 3.1.46 - 09/15/23
 -----------------